Remove duplicate models
