+++ src/main/java/com/saucelabs/saucerest/model/storage/DeleteAppFile.java
@@ -1,5 +1,5 @@
 
-package com.saucelabs.saucerest.model.storage.deleteappfile;
+package com.saucelabs.saucerest.model.storage;
 
 import com.squareup.moshi.Json;
