+++ src/main/java/com/saucelabs/saucerest/model/storage/EditFileDescription.java
@@ -1,5 +1,5 @@
 
-package com.saucelabs.saucerest.model.storage.editfiledescription;
+package com.saucelabs.saucerest.model.storage;
 
 import com.squareup.moshi.Json;
