+++ src/main/java/com/saucelabs/saucerest/model/storage/Item.java
@@ -1,9 +1,6 @@
 
-package com.saucelabs.saucerest.model.storage.deleteappfile;
+package com.saucelabs.saucerest.model.storage;
 
-import com.saucelabs.saucerest.model.storage.Access;
-import com.saucelabs.saucerest.model.storage.Metadata;
-import com.saucelabs.saucerest.model.storage.Owner;
 import com.squareup.moshi.Json;
 
 public class Item {
