+++ src/main/java/com/saucelabs/saucerest/model/storage/UploadFileApp.java
@@ -1,5 +1,5 @@
 
-package com.saucelabs.saucerest.model.storage.uploadfileapp;
+package com.saucelabs.saucerest.model.storage;
 
 import com.squareup.moshi.Json;
