+++ src/test/java/com/saucelabs/saucerest/integration/StorageTest.java
@@ -4,13 +4,9 @@ import com.google.common.collect.ImmutableMap;
 import com.saucelabs.saucerest.DataCenter;
 import com.saucelabs.saucerest.SauceREST;
 import com.saucelabs.saucerest.api.Storage;
-import com.saucelabs.saucerest.model.storage.deleteappfile.DeleteAppFile;
+import com.saucelabs.saucerest.model.storage.*;
 import com.saucelabs.saucerest.model.storage.deletegroupappfiles.DeleteAppGroupFiles;
-import com.saucelabs.saucerest.model.storage.editappgroupsettings.EditAppGroupSettings;
-import com.saucelabs.saucerest.model.storage.getappfiles.GetAppFiles;
 import com.saucelabs.saucerest.model.storage.getappgroups.GetAppStorageGroups;
-import com.saucelabs.saucerest.model.storage.getappgroupsettings.GetAppStorageGroupSettings;
-import com.saucelabs.saucerest.model.storage.uploadfileapp.UploadFileApp;
 import org.awaitility.Awaitility;
 import org.json.JSONObject;
 import org.junit.jupiter.api.Assertions;
